+++ app/components/notification_banner.rb
@@ -1,5 +1,6 @@
 class NotificationBanner < ViewComponent::Base
-  def initialize(flashes:)
+  def initialize(flashes: nil, message: nil)
     @flashes = flashes
+    @message = message
   end
 end